Bulk test fixes
