+++ src/engine/Engine.h
@@ -174,17 +174,6 @@ public:
     void setVerbosity( unsigned verbosity );
 
     /*
-      Apply the stack to the newly created SmtCore, returns false if UNSAT is
-      found in this process.
-    */
-    bool restoreSmtState( SmtState &smtState );
-
-    /*
-      Store the current stack of the smtCore into smtState
-    */
-    void storeSmtState( SmtState &smtState );
-
-    /*
       Pick the piecewise linear constraint for splitting
     */
     PiecewiseLinearConstraint *pickSplitPLConstraint( DivideStrategy strategy );
