+++ src/engine/SmtCore.h
@@ -2,55 +2,96 @@
 /*! \file SmtCore.h
  ** \verbatim
  ** Top contributors (to current version):
- **   Guy Katz, Parth Shah
+ **   Guy Katz, AleksandarZeljic, Haoze Wu, Parth Shah
  ** This file is part of the Marabou project.
  ** Copyright (c) 2017-2019 by the authors listed in the file AUTHORS
  ** in the top-level source directory) and their institutional affiliations.
  ** All rights reserved. See the file COPYING in the top-level source
  ** directory for licensing information.\endverbatim
  **
- ** [[ Add lengthier description here ]]
-
-**/
+ ** The SmtCore class implements a context-dependent SmtCore class.
+ ** The SmtCore distinguishes between: **decisions** and **implications**.
+ **
+ ** Decision is a case of PiecewiseLinearConstraint asserted on the trail.
+ ** A decision is a choice between multiple feasible cases of a
+ ** PiecewiseLinearConstraint and represents nodes in the search tree.
+ **
+ ** Implication is the last feasible case of a PiecewiseLinearConstraint and as
+ ** soon as it is detected it is asserted on the trail.
+ **
+ ** Case splitting on a PiecewiseLinearConstraint performs a decision.
+ ** Fixing a case of a PiecewiseLinearConstraint (e.g., via bound propagation or
+ ** by exhausting all other cases) performs an implication.
+ **
+ ** The overall search state is stored in a distributed way:
+ ** - SmtCore::_trail is the current search state in a chronological order
+ ** - PiecewiseLinearConstraints' infeasible cases enumerate all the explored
+ ** states w.r.t to the chronological order on the _trail.
+ **
+ ** _trail is a chronological list of cases of PiecewiseLinearConstraints
+ ** asserted to hold (as TrailEntries) - both decisions and implications.
+ ** _decisions is a chronological list of decisions stored on the trail.
+ ** _trail and _decisions are both context dependent and will synchronize in
+ ** unison with the _context object.
+ **
+ ** When a search state is found to be infeasible, SmtCore backtracks to the
+ ** last decision and continues the search.
+ **
+ ** PushDecision advances the decision level and context level.
+ ** popDecisionLevel backtracks the last decision and context level.
+ **
+ ** Advancing a context level creates a synchronization point to which all
+ ** context dependent members can backtrack in unison.
+ ** Backtracking a context level restores all context dependent members to the
+ ** state of the last synchronization point. This operation is O(1) complexity.
+ **
+ ** Since the entire search state is context dependent, backtracking the context
+ ** (via popDecisionLevel) backtracks the entire Marabou search state.
+ ** The only exception is the labeling of basic/non-basic variables in the
+ ** tableau, which may need to be recalculated using Tableau's
+ ** postContextPopHook exposed via Engine.
+ **
+ ** Implementation relies on:
+ **
+ ** - _context is a unique Context object from which all the context-dependent
+ ** structures are obtained.
+ **
+ ** - PiecewiseLinearConstraint class stores its search state in a
+ ** context-dependent manner and exposes it using nextFeasibleCase() and
+ ** markInfeasible() methods.
+ **
+ ** - Using BoundManager class to store bounds in a context-dependent manner
+ **/
 
 #ifndef __SmtCore_h__
 #define __SmtCore_h__
 
-#include "DivideStrategy.h"
+#include "Options.h"
 #include "PiecewiseLinearCaseSplit.h"
 #include "PiecewiseLinearConstraint.h"
 #include "PLConstraintScoreTracker.h"
-#include "SmtState.h"
 #include "Stack.h"
-#include "SmtStackEntry.h"
 #include "Statistics.h"
+#include "TrailEntry.h"
+#include "context/cdlist.h"
 #include "context/context.h"
-
-#include <memory>
 
 #define SMT_LOG( x, ... ) LOG( GlobalConfiguration::SMT_CORE_LOGGING, "SmtCore: %s\n", x )
 
 class EngineState;
-class IEngine;
+class Engine;
 class String;
-
-using CVC4::context::Context;
 
 class SmtCore
 {
 public:
-    SmtCore( IEngine *engine );
+    SmtCore( IEngine *engine, CVC4::context::Context &context );
     ~SmtCore();
 
     /*
       Clear the stack.
     */
     void freeMemory();
-
-    /*
-      Reset the SmtCore
-    */
-    void reset();
 
     /*
       Initialize the score tracker with the given list of pl constraints.
@@ -90,13 +131,12 @@ public:
       Get the number of times a specific PL constraint has been reported as
       violated.
     */
-    unsigned getViolationCounts( PiecewiseLinearConstraint* constraint ) const;
+    unsigned getViolationCounts( PiecewiseLinearConstraint *constraint ) const;
 
     /*
-      Reset all reported violation counts and the number of rejected SoI
-      phase pattern proposal.
+      Reset all reported violation counts.
     */
-    void resetSplitConditions();
+    void resetReportedViolations();
 
     /*
       Returns true iff the SMT core wants to perform a case split.
@@ -104,32 +144,80 @@ public:
     bool needToSplit() const;
 
     /*
-      Perform the split according to the constraint marked for
-      splitting. Update bounds, add equations and update the stack.
-    */
-    void performSplit();
+       Push TrailEntry representing the decision onto the trail.
+     */
+    void pushDecision( PiecewiseLinearConstraint *constraint,  PhaseStatus decision );
 
     /*
-      Pop an old split from the stack, and perform a new split as
-      needed. Return true if successful, false if the stack is empty.
+      Inform SmtCore of an implied (formerly valid) case split that was discovered.
+    */
+    void pushImplication( PiecewiseLinearConstraint *constraint );
+
+    /*
+        Pushes trail entry onto trail, handles decision book-keeping and
+        update bounds and add equations to the engine.
+     */
+    void applyTrailEntry( TrailEntry &te, bool isDecision = false );
+
+    /*
+      Decide and apply a case split using the constraint marked for splitting.
+    */
+    void decide();
+
+    /*
+      Decide a constraint's feasible case. Update bounds, add equations
+      and trail.
+     */
+    void decideSplit( PiecewiseLinearConstraint *constraint );
+
+    /*
+       Backtracks fully explored decisions and stores the last feasible decision
+     */
+    bool backtrackToFeasibleDecision( TrailEntry &feasibleDecision );
+
+    /*
+      Return to a feasible state and resume search by asserting the next case
+      (as either a decision or implication)
+    */
+    bool backtrackAndContinueSearch();
+
+    /*
+      Pop a stack frame. Return true if successful, false if the stack is empty.
     */
     bool popSplit();
 
     /*
+      Pop a context level - lazily backtracking trail, bounds, etc.
+      Return true if successful, false if the stack is empty.
+    */
+    bool popDecisionLevel( TrailEntry &lastDecision );
+
+    /*
       The current stack depth.
     */
-    unsigned getStackDepth() const;
+    unsigned getDecisionLevel() const;
 
     /*
-      Let the smt core know of an implied valid case split that was discovered.
-    */
-    void recordImpliedValidSplit( PiecewiseLinearCaseSplit &validSplit );
-
-    /*
-      Return a list of all splits performed so far, both SMT-originating and valid ones,
-      in the correct order.
+      Return a list of all splits performed so far, both SMT-originating and
+      valid ones, in the correct order.
     */
     void allSplitsSoFar( List<PiecewiseLinearCaseSplit> &result ) const;
+
+    /*
+      Get trail begin iterator.
+    */
+    CVC4::context::CDList<TrailEntry>::const_iterator trailBegin() const
+    {
+        return _trail.begin();
+    };
+
+    /*
+      Get trail end iterator.
+    */
+    CVC4::context::CDList<TrailEntry>::const_iterator trailEnd() const
+    {
+        return _trail.end();
+    };
 
     /*
       Have the SMT core start reporting statistics.
@@ -142,20 +230,12 @@ public:
     */
     PiecewiseLinearConstraint *chooseViolatedConstraintForFixing( List<PiecewiseLinearConstraint *> &_violatedPlConstraints ) const;
 
+    void setConstraintViolationThreshold( unsigned threshold );
+
     inline void setBranchingHeuristics( DivideStrategy strategy )
     {
         _branchingHeuristic = strategy;
     }
-
-    /*
-      Replay a stackEntry
-    */
-    void replaySmtStackEntry( SmtStackEntry *stackEntry );
-
-    /*
-      Store the current state of the SmtCore into smtState
-    */
-    void storeSmtState( SmtState &smtState );
 
     /*
       Pick the piecewise linear constraint for splitting, returns true
@@ -164,37 +244,46 @@ public:
     bool pickSplitPLConstraint();
 
     /*
+     * For testing purposes
+     */
+    PiecewiseLinearCaseSplit getDecision( unsigned decisionLevel ) const;
+
+    void reset();
+
+    /*
       For debugging purposes only - store a correct possible solution
+      TODO: Create a user interface for this
     */
     void storeDebuggingSolution( const Map<unsigned, double> &debuggingSolution );
     bool checkSkewFromDebuggingSolution();
     bool splitAllowsStoredSolution( const PiecewiseLinearCaseSplit &split, String &error ) const;
-
+    void interruptIfCompliantWithDebugSolution();
 private:
-    /*
-      Valid splits that were implied by level 0 of the stack.
-    */
-    List<PiecewiseLinearCaseSplit> _impliedValidSplitsAtRoot;
-
     /*
       Collect and print various statistics.
     */
     Statistics *_statistics;
 
     /*
-      The case-split stack.
+      CVC4 Context, constructed in Engine
     */
-    List<SmtStackEntry *> _stack;
+    CVC4::context::Context &_context;
+
+    /*
+      Trail is context dependent and contains all the asserted PWLCaseSplits
+    */
+    CVC4::context::CDList<TrailEntry> _trail;
+
+    /*
+     * _decisions stores the decision TrailEntries
+     */
+    CVC4::context::CDList<TrailEntry> _decisions;
 
     /*
       The engine.
     */
     IEngine *_engine;
 
-    /*
-      Context for synchronizing the search.
-     */
-    Context &_context;
     /*
       Do we need to perform a split and on which constraint.
     */
@@ -210,12 +299,6 @@ private:
       For debugging purposes only
     */
     Map<unsigned, double> _debuggingSolution;
-
-    /*
-      A unique ID allocated to every state that is stored, for
-      debugging purposes.
-    */
-    unsigned _stateId;
 
     /*
       Split when some relu has been violated for this many times during the
@@ -244,14 +327,7 @@ private:
       current search state.
     */
     unsigned _numRejectedPhasePatternProposal;
+
 };
 
 #endif // __SmtCore_h__
-
-//
-// Local Variables:
-// compile-command: "make -C ../.. "
-// tags-file-name: "../../TAGS"
-// c-basic-offset: 4
-// End:
-//
